+++ lib/shoko_ui.dart
@@ -5,5 +5,6 @@ export 'src/styles/shadow.dart';
 export 'src/widgets/button.dart';
 export 'src/widgets/switch.dart';
 export 'src/widgets/radio.dart';
-export 'src/widgets/textfield.dart';
+export 'src/widgets/textfields/textfield.dart';
+export 'src/widgets/textfields/modern_textfield.dart';
 export 'src/theme/shoko_ui_theme.dart';