+++ lib/src/widgets/button.dart
@@ -28,7 +28,7 @@ class ShokoUIButton extends StatelessWidget {
     this.onTap, this.onDoubleTap, this.onHover, this.onLongPress,
     this.color, this.isOutline = false, this.isFullWidth = false, this.isChildAtCenter = false,
     required this.child, this.suffix, this.prefix,
-    this.radius = ShokoUIRadii.medium, this.shadow
+    this.radius, this.shadow
   });
 
   @override
@@ -41,6 +41,8 @@ class ShokoUIButton extends StatelessWidget {
         width: isFullWidth ? double.maxFinite : null,
         padding: const EdgeInsets.symmetric(vertical: 16, horizontal: 16),
         decoration: BoxDecoration(
+          color: color ?? (context.shokoTheme?.buttonTheme?.color ?? Colors.grey[850]),
+          borderRadius: (radius?.get() ?? context.shokoTheme?.buttonTheme?.radius?.get()) ?? ShokoUIRadii.medium.get(),
           color: !isOutline ? color ?? (context.shokoTheme?.buttonTheme?.color ?? Colors.grey[850]) : null,
           borderRadius: (radius?.get() ?? context.shokoTheme?.buttonTheme?.radius?.get()),
           boxShadow: (shadow?.get() ?? context.shokoTheme?.buttonTheme?.shadow?.get()),