+++ lib/src/widgets/textfields/modern_textfield.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:shoko_ui/shoko_ui.dart';
 
-class ShokoUITextField extends StatefulWidget {
+class ShokoUIModernTextField extends StatefulWidget {
   
   ///Be auto-dispose when widget disposed
   final TextEditingController controller;
@@ -10,14 +10,12 @@ class ShokoUITextField extends StatefulWidget {
   final Function(String value)? onChange;
   final bool Function(String value)? validator;
 
-  final EdgeInsets? insidePaddings;
-  final ShokoUIRadii? radius;
-
   final bool isEnabled;
+  final bool obscureText;
 
   final bool isError;
   final String? errorText;
-  final Widget? errorWidget;
+  final TextStyle? errorTextStyle;
 
   final String? label;
   final TextStyle? labelTextStyle;
@@ -27,29 +25,24 @@ class ShokoUITextField extends StatefulWidget {
   final Color? focusColor;
   final Color? errorColor;
 
-  final double? borderWidth;
-  final double? width;
-
-  final TextInputType? keyboardType;
-
+  final int? minSymbols;
+  final int? maxSymbols;
   
-  const ShokoUITextField({super.key,
+  const ShokoUIModernTextField({super.key,
     required this.controller, this.style,
     this.onChange, this.validator,
-    this.insidePaddings, this.radius = ShokoUIRadii.medium,
     this.isEnabled = true,
-    this.isError = false, this.errorText, this.errorWidget,
+    this.isError = false, this.errorText, this.errorTextStyle,
     this.label, this.labelTextStyle,
     this.enableColor, this.disableColor, this.focusColor, this.errorColor,
-    this.borderWidth, this.width,
-    this.keyboardType
+    this.maxSymbols, this.minSymbols, this.obscureText = false
   });
 
   @override
-  State<ShokoUITextField> createState() => _ShokoUITextFieldState();
+  State<ShokoUIModernTextField> createState() => _ShokoUIModernTextFieldState();
 }
 
-class _ShokoUITextFieldState extends State<ShokoUITextField> {
+class _ShokoUIModernTextFieldState extends State<ShokoUIModernTextField> {
 
   final FocusNode focusNode = FocusNode();
   bool isError = false;
@@ -62,7 +55,7 @@ class _ShokoUITextFieldState extends State<ShokoUITextField> {
   }
 
   @override
-  void didUpdateWidget(covariant ShokoUITextField oldWidget) {
+  void didUpdateWidget(covariant ShokoUIModernTextField oldWidget) {
     setState(() {
       isError = widget.isError;
     });
@@ -78,7 +71,7 @@ class _ShokoUITextFieldState extends State<ShokoUITextField> {
 
   submit(String value) {
     if(widget.onChange != null) widget.onChange!(value);
-            
+    
     bool validatorResult = true;
     if (widget.validator != null) {
       validatorResult = widget.validator!(value);
@@ -97,7 +90,6 @@ class _ShokoUITextFieldState extends State<ShokoUITextField> {
       crossAxisAlignment: CrossAxisAlignment.start,
       mainAxisSize: MainAxisSize.min,
       children: [
-        if(widget.label != null) Text(widget.label!, style: (widget.labelTextStyle ?? theme?.labelTextStyle) ?? const TextStyle()),
         Focus(
           onFocusChange: (value) {
             !value ? submit(this.value) : setState(() {});
@@ -105,31 +97,38 @@ class _ShokoUITextFieldState extends State<ShokoUITextField> {
           child: AnimatedContainer(
             duration: const Duration(milliseconds: 500),
             curve: Curves.ease,
-            width: (widget.width ?? theme?.width),
-            padding: (widget.insidePaddings ?? theme?.insidePaddings) ?? const EdgeInsets.all(12),
             decoration: BoxDecoration(
-              borderRadius: (widget.radius?.get() ?? context.shokoTheme?.buttonTheme?.radius?.get()),
+              borderRadius: ShokoUIRadii.mediumPlus.get(),
               border: Border.all(
-                width: (widget.borderWidth ?? theme?.borderWidth) ?? 2,
-                color: (focusNode.hasFocus && widget.isEnabled ? (widget.focusColor ?? theme?.focusColor) : (widget.isEnabled ? (isError ? (widget.errorColor ?? theme?.errorColor) : (widget.enableColor ?? theme?.enableColor)) : (widget.disableColor ?? theme?.disableColor))) ??Colors.grey[850]!
+                width: 2,
+                color: (focusNode.hasFocus && widget.isEnabled ? (widget.focusColor ?? theme?.focusColor) : (widget.isEnabled ? (isError ? (widget.errorColor ?? theme?.errorColor) : (widget.enableColor ?? theme?.enableColor)) : (widget.disableColor ?? theme?.disableColor)) ?? Colors.black)!
               )
             ),
-            child: EditableText(
-              keyboardType: widget.keyboardType,
+            child: TextField(
               onTapOutside: (event) => focusNode.unfocus(),
               onSubmitted: submit,
-              // onEditingComplete: () => submit(value),
               onChanged: (value) => this.value = value,
               readOnly: !widget.isEnabled,
               controller: widget.controller,
               focusNode: focusNode,
-              style: ((widget.style ?? theme?.style) ?? const TextStyle(color: Colors.black, fontSize: 16)).copyWith(color: !widget.isEnabled ? ((widget.disableColor ?? theme?.disableColor) ?? Colors.grey) : null),
-              cursorColor: Colors.black,
-              backgroundCursorColor: Colors.black
+              maxLength: widget.maxSymbols,
+              obscureText: widget.obscureText,
+              enableSuggestions: true,
+              autocorrect: true,
+              maxLines: 1,
+              decoration: InputDecoration(
+                enabled: widget.isEnabled,
+                counterText: '',
+                contentPadding: const EdgeInsets.only(left: 16, top: 8, bottom: 8),
+                label: (widget.label != null) ? Text(widget.label!) : null,
+                labelStyle: (widget.labelTextStyle ?? theme?.labelTextStyle ?? TextStyle(color: focusNode.hasFocus ? widget.focusColor : (widget.isEnabled ? isError ? widget.errorColor ?? theme?.errorColor: (widget.enableColor ?? theme?.enableColor) : widget.disableColor ?? theme?.disableColor ?? Colors.grey[400]))),
+                border: InputBorder.none
+              ),
+              style: ((widget.style ?? theme?.style) ?? const TextStyle(color: Colors.black, fontSize: 16)).copyWith(color: !widget.isEnabled ? (widget.disableColor ?? theme?.disableColor ?? Colors.grey[400]) : null),
             )
           )
         ),
-        if (isError && (widget.errorWidget != null || widget.errorText != null)) (widget.errorWidget ?? Text(widget.errorText!, style: TextStyle(color: widget.errorColor ?? Colors.red),))
+        if (isError && widget.errorText != null) (Text(widget.errorText!, style: widget.errorTextStyle ?? theme?.errorTextStyle ?? TextStyle(color: widget.errorColor ?? Colors.red),))
       ]
     );
   }